# docs/openapi.yaml
openapi: 3.0.3
info:
  title: MCP Tools API
  description: |
    RESTful API for the MCP Tools ecosystem providing unified access to
    Kanban boards, Memory graph, Wiki pages, Calendar events, and Monitoring.
  version: 1.0.0
  contact:
    name: MCP Tools Team
    url: https://github.com/your-org/mcp-tools
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.mcp-tools.dev/v1
    description: Production server
  - url: http://localhost:3001/v1
    description: Local development server
security:
  - bearerAuth: []
paths:
...
  /kanban/boards:
    get:
      tags: [Kanban]
      summary: List boards
      parameters:
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/limit'
        - name: sort
          in: query
          schema:
            type: string
            enum: [name, created_at, updated_at]
            default: created_at
        - name: order
          in: query
          schema:
            type: string
            enum: [asc, desc]
            default: desc
      responses:
        '200':
          description: List of boards
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/BoardSummary'
                      pagination:
                        $ref: '#/components/schemas/Pagination'
    post:
      tags: [Kanban]
      summary: Create board
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateBoardRequest'
      responses:
        '201':
          description: Board created
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Board'
  /kanban/boards/{boardId}:
    get:
      tags: [Kanban]
      summary: Get board with columns and cards
      parameters:
        - name: boardId
          in: path
          required: true
          schema:
            type: string
        - name: include_archived
          in: query
          schema:
            type: boolean
            default: false
      responses:
        '200':
          description: Board details
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/BoardWithColumns'
  /kanban/cards:
    post:
      tags: [Kanban]
      summary: Create card
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCardRequest'
      responses:
        '201':
          description: Card created
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Card'
  /kanban/cards/{cardId}:
    put:
      tags: [Kanban]
      summary: Update card
      parameters:
        - name: cardId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCardRequest'
      responses:
        '200':
          description: Card updated
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Card'
    delete:
      tags: [Kanban]
      summary: Delete card
      parameters:
        - name: cardId
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Card deleted
  /kanban/cards/{cardId}/move:
    put:
      tags: [Kanban]
      summary: Move card
      parameters:
        - name: cardId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MoveCardRequest'
      responses:
        '200':
          description: Card moved
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Card'

  # Memory API
...
  /wiki/pages:
    get:
      tags: [Wiki]
      summary: List wiki pages
      parameters:
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/limit'
        - name: category_id
          in: query
          schema:
            type: string
        - name: tags
          in: query
          schema:
            type: array
            items:
              type: string
          style: form
          explode: true
        - name: search
          in: query
          schema:
            type: string
        - name: sort
          in: query
          schema:
            type: string
            enum: [title, created_at, updated_at]
            default: updated_at
      responses:
        '200':
          description: List of wiki pages
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/WikiPageSummary'
                      pagination:
                        $ref: '#/components/schemas/Pagination'
    post:
      tags: [Wiki]
      summary: Create wiki page
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateWikiPageRequest'
      responses:
        '201':
          description: Wiki page created
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/WikiPage'
  /wiki/pages/{pageId}:
    get:
      tags: [Wiki]
      summary: Get wiki page
      parameters:
        - name: pageId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Wiki page details
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/WikiPage'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
  schemas:
    SuccessResponse:
      type: object
      properties:
        meta:
          type: object
          properties:
            timestamp:
              type: string
              format: date-time
            requestId:
              type: string
          required: [timestamp, requestId]
      required: [meta]
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        hasNext:
          type: boolean
        hasPrev:
          type: boolean
      required: [page, limit, total, hasNext, hasPrev]

    # Kanban Schemas
    BoardSummary:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        color:
          type: string
          pattern: '^#[0-9A-Fa-f]{6}$'
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        columns_count:
          type: integer
        cards_count:
          type: integer
      required: [id, name, color, created_at, updated_at, columns_count, cards_count]
    Board:
      allOf:
        - $ref: '#/components/schemas/BoardSummary'
    BoardWithColumns:
      allOf:
        - $ref: '#/components/schemas/Board'
        - type: object
          properties:
            columns:
              type: array
              items:
                $ref: '#/components/schemas/ColumnWithCards'
    ColumnWithCards:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        color:
          type: string
          pattern: '^#[0-9A-Fa-f]{6}$'
        position:
          type: integer
        cards:
          type: array
          items:
            $ref: '#/components/schemas/Card'
      required: [id, name, color, position, cards]
    Card:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: string
        priority:
          type: string
          enum: [low, medium, high, urgent]
        assigned_to:
          type: string
        tags:
          type: array
          items:
            type: string
        due_date:
          type: string
          format: date-time
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        position:
          type: integer
      required: [id, title, priority, created_at, updated_at, position]
    CreateBoardRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 255
        description:
          type: string
        color:
          type: string
          pattern: '^#[0-9A-Fa-f]{6}$'
          default: '#6366f1'
      required: [name]
    CreateCardRequest:
      type: object
      properties:
        board_id:
          type: string
        column_id:
          type: string
        title:
          type: string
          minLength: 1
          maxLength: 255
        description:
          type: string
        priority:
          type: string
          enum: [low, medium, high, urgent]
          default: medium
        assigned_to:
          type: string
        tags:
          type: array
          items:
            type: string
        due_date:
          type: string
          format: date-time
        position:
          type: integer
          default: 0
      required: [board_id, column_id, title]
    UpdateCardRequest:
      type: object
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 255
        description:
          type: string
        priority:
          type: string
          enum: [low, medium, high, urgent]
        assigned_to:
          type: string
        tags:
          type: array
          items:
            type: string
        due_date:
          type: string
          format: date-time
    MoveCardRequest:
      type: object
      properties:
        column_id:
          type: string
        position:
          type: integer
          minimum: 0
      required: [column_id, position]

    # Memory Schemas
...
    WikiPageSummary:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        slug:
          type: string
        summary:
          type: string
        category:
          type: object
          properties:
            id:
              type: string
            name:
              type: string
            slug:
              type: string
          required: [id, name, slug]
        tags:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
              name:
                type: string
              color:
                type: string
            required: [id, name, color]
        author_id:
          type: string
        published:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        version:
          type: integer
      required: [id, title, slug, author_id, published, created_at, updated_at, version]
    WikiPage:
      allOf:
        - $ref: '#/components/schemas/WikiPageSummary'
        - type: object
          properties:
            content:
              type: string
            html_content:
              type: string
            table_of_contents:
              type: array
              items:
                type: object
                properties:
                  level:
                    type: integer
                  title:
                    type: string
                  anchor:
                    type: string
                required: [level, title, anchor]
          required: [content]
    CreateWikiPageRequest:
      type: object
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 255
        content:
          type: string
          minLength: 1
        category_id:
          type: string
        tags:
          type: array
          items:
            type: string
        published:
          type: boolean
          default: true
      required: [title, content]